Okta login should only be valid for a given domain
We should only trust Okta to authenticate a user within the domain we've
registered the SSO metadata for.

An admin within Okta can change a user's email to be whatever they want,
including that of someone in a different organization. If we didn't have
this check, we would be allowing them to login as any other user within
our sytem.