Upgrade to ES 2.0
AWS ES Service only supports version 1.5 of ES. It has limited
functionality which prevents us from using it (can't open and close
indexes, which makes adding types to an index require
deleting and recreating the index). We're going to use elastic's "found"
product, which is hosted ES and supports version 2.0.