Use build-requirements and wheel-requirements
pip doesn't handle creating and installing from wheels correctly with
`--no-index` if the requirement is specified with a `git+`.

previously, with just requirements.txt, anytime we built the image, we
had to clone the repos to build the wheel directory, and then again when
installing from the wheel directory.

by using the package names, we can leverage the wheels that are built.